agentverse-aura/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aura-agentverse"
version = "0.1.0"
description = "AURA Smart Home Management Agent for FetchAI AgentVerse"
authors = [
    {name = "AURA Team", email = "aura@example.com"},
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "agentverse>=0.2.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']
